More permissions: admin user and regular users. More tests

+++ views.py
@@ -39,6 +39,7 @@ def login_required(test):
 def logout():
     session.pop('logged_in', None)
     session.pop('user_id', None)
+    session.pop('role', None)
     flash('You are logged out. Bye.')
     return redirect(url_for('login'))
 
@@ -62,6 +63,7 @@ def login():
             else:
                 session['logged_in'] = True
                 session['user_id'] = u.id
+                session['role'] = u.role
                 flash('You are logged in. Go crazy.')
                 return redirect(url_for('tasks'))
         else:
@@ -145,7 +147,7 @@ def new_task():
 def complete(task_id):
     new_id = task_id
     task = db.session.query(Task).filter_by(task_id=new_id)
-    if session['user_id'] == task.first().user_id:
+    if session['user_id'] == task.first().user_id or session['role'] == "admin":
         task.update({"status":"0"})
         db.session.commit()
         flash('The task was marked as complete. Nice.')
@@ -158,9 +160,13 @@ def complete(task_id):
 @login_required
 def delete_entry(task_id):
     new_id = task_id
-    db.session.query(Task).filter_by(task_id=new_id).delete()
-    db.session.commit()
-    flash('The task was deleted. Why not add a new one?')
+    task = db.session.query(Task).filter_by(task_id=new_id)
+    if session['user_id'] == task.first().user_id or session['role'] == "admin":
+        db.session.query(Task).filter_by(task_id=new_id).delete()
+        db.session.commit()
+        flash('The task was deleted. Why not add a new one?')
+    else:
+        flash('That task belongs to another user.')
     return redirect(url_for('tasks'))
 
 def flash_errors(form):
